requirements.txt
Pillow
pypng
pyqrcode
PyQt5
phonenumbers
validate_email
password_strength